requirements.txt
absl-py==1.4.0
cached-property==1.5.2
certifi==2021.5.30
charset-normalizer==2.0.12
cvzone==1.5.6
cycler==0.11.0
Cython==0.29.33
dataclasses @ file:///tmp/build/80754af9/dataclasses_1614363715916/work
docopt==0.6.2
h5py==3.1.0
hdf5storage==0.1.18
idna==3.4
kiwisolver==1.3.1
matplotlib==3.3.4
mediapipe==0.8.2
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy==1.19.3
olefile==0.46
opencv-python==3.4.13.47
pandas==1.1.5
Pillow==8.4.0
pipreqs==0.4.11
protobuf==3.19.6
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
requests==2.27.1
scipy==1.5.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
tensorboardX==2.5.1
torch==1.10.2
torchaudio==0.10.2
torchvision==0.11.3
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3==1.26.15
wincertstore==0.2
yacs==0.1.8
yarg==0.1.9
